All Workinggit add .-S

+++ test6.c
@@ -2,9 +2,9 @@ int n=10;
 for(int i=0;i<n;i++)
 {
     int loop_var =0;
-    int level1_var =0;
+    float level1_var =0;
     while (counter < 5) {
-        int loop_var = 3 * counter; 
+        float loop_var = 3 * counter; 
         counter++;
 
         if (loop_var > level1_var) {
